+++ opm/models/blackoil/blackoilfoammodules.hh
@@ -36,10 +36,10 @@
 //#include <opm/material/common/IntervalTabulated2DFunction.hpp>
 
 #if HAVE_ECL_INPUT
-#include <opm/parser/eclipse/Deck/Deck.hpp>
-#include <opm/parser/eclipse/EclipseState/EclipseState.hpp>
-#include <opm/parser/eclipse/EclipseState/Tables/FoamadsTable.hpp>
-#include <opm/parser/eclipse/EclipseState/Tables/FoammobTable.hpp>
+#include <opm/input/eclipse/Deck/Deck.hpp>
+#include <opm/input/eclipse/EclipseState/EclipseState.hpp>
+#include <opm/input/eclipse/EclipseState/Tables/FoamadsTable.hpp>
+#include <opm/input/eclipse/EclipseState/Tables/FoammobTable.hpp>
 #endif
 
 #include <opm/material/common/Valgrind.hpp>
